+++ misosoup/library/common.py
@@ -1,7 +1,6 @@
 """Common functions."""
 import re
 import yaml
-from tqdm import tqdm
 
 
 def get_metabolite_name(compound):
@@ -41,7 +40,7 @@ def merge_dicts(a, b, path=None):
 def merge_yaml(file_list):
     """Merge yaml files."""
     solutions = {}
-    for path in tqdm(file_list):
+    for path in file_list:
         with open(path, "r") as file_descriptor:
             merge_dicts(solutions, yaml.safe_load(file_descriptor))
     return solutions